Extract HomeCarousel into generic component
